test(exports): consolidate fixtures into fixtures.py and repoint imports
